duplicated 'eval' related functions as 'worker' ones.
This is in preparation for converting the eval operation to behave as
it was defined in the original spec.

Currently, both 'eval' and 'worker' start a new worker process.

In future the 'eval' operation will act like an 'out' operation, where
the elements of the tuple will be evaluated in a separate worker process,
and then added to the tuple space.