tspace ETS table now uses integer for key.
The integer() key replaces the reference() key. This speeds up addition of
tuples, since we just increment a number, rather than calling make_ref().

We keep track of the next key (number) to use in the server state.

The table type has also been changed to ordered_set, which should ensure
that if two or more tuples match in an "in" or "rd" operation, we return
the first one that was inserted.