terminating espace server sends `quit' to clients.
Before this change the blocking clients, upon termination of the server,
would wait forever. Now, they will receive the 'quit' atom. This only
affects the in/1,2 and rd/1,2 functions.

In most cases this will probably result in a 'badmatch' exception.

However, in the cases where the client assumes that return from the
function is an indication of presene of tuple, then it may/will cause
problems!