performance improvement for named instances.
For named instances, we rely on espace_util:inst_to_name/2 to generate
a suitable Prefix_Instname atom. For repeated operations this is costly,
so we generate the concatenated string once and cache it in the process
dictionary.

This works well for processes (workers) that make many repeated espace
calls, however, it does not help if we have many short lived operations,
such as evals!