remove the optimization for "rd"
Unlike "rdp", which either finds and returns the tuple or it returns
"nomatch", if "rd" does not find a match, it will need to block via
espace_tspatt_srv. If a tuple is added by another process between
these two steps, the original process calling "rd" may never be
notified!